rename: 'numeric' to 'quantitative'
