+++ src/classes/CoordinateTree/CoordinateTransformation.js
@@ -54,10 +54,10 @@ export default class CoordinateTransformation {
     // For categorical and temporal domains, we don't need to apply the scaling,
     // since we've already done this when the prop was passed (see Mark.js,
     // parseCoord function). This is because we need to support nested Sections,
-    // where the parent is for example categorical but the child is numeric. It
+    // where the parent is for example categorical but the child is quantitative. It
     // is also necessary to properly interpolate in the interpolatePath function
     // (you cannot interpolate between 'A' and 'B'). So for these components,
-    // we already need to know the converted (numeric) value before the transform
+    // we already need to know the converted (quantitative) value before the transform
     // function is used.
     this.getX = x => {
       if (['categorical', 'temporal'].includes(this.domainTypes.x)) {
