+++ src/scales/shorthands/color/quantitative.js
