+++ src/scales/shorthands/coords/quantitative.js
