+++ src/scales/shorthands/opacity/quantitative.js
